+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Fix: Checkout fields validation.
     * Fix: Checkout registration.
     * Fix: Remove usage of `price` meta from purchasable products.
+    * Fix: Properly handle account creation errors on checkout.
 * 2.0.5 - 2016.11.24:
     * Fix: Updating customer on checkout page.
     * Fix: Removed unnecessary filter.